massive re-work of library and API
biggest change is that all objects are now fully contained and use accessor
funcs to get at properites. should be a much cleaner interface now.

also, the default is now to flush/write to _all_ enabled backends rather than
leaving it up to the user to manage which backends should be written to.