Don't bother sending an empty message to Kafka
There was previously a corner case where a KP of just the right size could cause TSK to send an empty message if the last key in the KP triggered a send based on the `SEND_IF_FULL` check. Then when the loop exits, the `SEND_MSG` would send an empty buffer. This isn't really a problem since the client can handle an empty message, we might as well avoid sending it in the first place.