Minor improvements to kafka backend and tsk-proxy
Mostly this avoids an incorrect log about the length of the Kafka message queue at shutdown time. Because we hadn't polled for Kafka callbacks recently, when we immediately query the queue length we get a stale response which makes it look like there are messages queued. This commit also significantly reduces the amount of time one must wait for a clean shutdown (~1s now compared to ~100s) after sending a SIGINT.