Add new kafka partitioning method: 'tskkey'
If the format is set to 'tskkey', data is assigned to a kafka
partition based on a hash of the series key (minus the last
term in the key, e.g. "pkt_cnt", "ip_len" etc for corsaro time
series data).

This means that all data points for a given series (and any
directly related series) will end up on the same partition. This
allows parallel kafka consumers in the same consumer group to be
guaranteed that they will see the full time series for any key
that they see, rather than it being spread across all consumers.

Default partitioning remains "tsk" for now.